backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
python-dotenv==1.2.2
google-generativeai==0.8.6
